typescript-sdk/integrations/openai-server/server/python/pyproject.toml
[tool.poetry]
name = "example_server"
version = "0.1.0"
description = ""
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ag-ui-protocol = "^0.1.5"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
openai = "^1.93.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "example_server:main"